Fix master memory leak due to MasterTaskExecuteRunnableDoesn't be removed
